services/api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
redis==5.0.4
prometheus-client==0.20.0
